python-server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-genai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.5.0
